build fixes for msvc2017, less warns

+++ core/buffers/runtimeTypeBuffer.cpp
@@ -238,7 +238,7 @@ G8Buffer *RuntimeTypeBuffer::toG8Buffer()
         {
             for (size_t j = 0; j < cols; j++)
             {
-                buffer->element(i, j) = (uint8_t)(at<float>(i, j));
+                buffer->element((int)i, (int)j) = (uint8_t)(at<float>(i, j));
             }
         }
     }
@@ -257,7 +257,7 @@ G12Buffer *RuntimeTypeBuffer::toG12Buffer(double min, double max)
         {
             for (size_t j = 0; j < cols; j++)
             {
-                buffer->element(i, j) = at<uint8_t>(i, j) << 4;
+                buffer->element((int)i, (int)j) = at<uint8_t>(i, j) << 4;
             }
         }
     }
@@ -269,7 +269,7 @@ G12Buffer *RuntimeTypeBuffer::toG12Buffer(double min, double max)
             {
                 int maxVal = G12Buffer::BUFFER_MAX_VALUE;
                 double value = lerpLimit(0, maxVal, at<float>(i, j), min, max);
-                buffer->element(i, j) = (uint16_t)value;
+                buffer->element((int)i, (int)j) = (uint16_t)value;
             }
         }
     }
@@ -288,7 +288,7 @@ RGB24Buffer *RuntimeTypeBuffer::toRGB24Buffer(double min, double max)
         {
             for (size_t j = 0; j < cols; j++)
             {
-                buffer->element(i, j) = RGBColor::gray(at<uint8_t>(i, j));
+                buffer->element((int)i, (int)j) = RGBColor::gray(at<uint8_t>(i, j));
             }
         }
     }
@@ -299,7 +299,7 @@ RGB24Buffer *RuntimeTypeBuffer::toRGB24Buffer(double min, double max)
             for (size_t j = 0; j < cols; j++)
             {
                 double value = lerpLimit(0, 0xFF, at<float>(i, j), min, max);
-                buffer->element(i, j) = RGBColor::gray(value);
+                buffer->element((int)i, (int)j) = RGBColor::gray(value);
             }
         }
     }
