+++ core/camerafixture/fixtureScene.cpp
@@ -526,7 +526,7 @@ bool FixtureScene::integrityRelink()
         FixtureCamera *cam = mOrphanCameras[i];
         cam->ownerScene = this;
         cam->cameraFixture = NULL;
-        cam->sequenceNumber = i;
+        cam->sequenceNumber = (int)i;
         for (size_t c = 0; c < cam->mImages.size(); c++)
         {
             ImageRelatedData *image = cam->mImages[c];
@@ -540,7 +540,7 @@ bool FixtureScene::integrityRelink()
     {
         CameraFixture *fixture = mFixtures[i];
         fixture->ownerScene = this;
-        fixture->sequenceNumber = i;
+        fixture->sequenceNumber = (int)i;
 
         vectorErase(fixture->cameras, (FixtureCamera *)NULL);
 
@@ -549,7 +549,7 @@ bool FixtureScene::integrityRelink()
             FixtureCamera *cam = fixture->cameras[j];
             cam->ownerScene = this;
             cam->cameraFixture = fixture;
-            cam->sequenceNumber = j;
+            cam->sequenceNumber = (int)j;
             for (size_t c = 0; c < cam->mImages.size(); c++)
             {
                 ImageRelatedData *image = cam->mImages[c];
